+++ app.py
@@ -15,20 +15,16 @@
 from __future__ import annotations
 
 import argparse
-from typing import NamedTuple, Union
 
 import dash
 import diskcache
-from dash import MATCH, DiskcacheManager, ctx
-from dash.dependencies import Input, Output, State
-from dash.exceptions import PreventUpdate
+from dash import DiskcacheManager
 
 from app_configs import APP_TITLE, THEME_COLOR, THEME_COLOR_SECONDARY
-from dash_html import generate_problem_details_table_rows, set_html
-from src.enums import SolverType
+from demo_interface import create_interface
 
-cache = diskcache.Cache("./cache")
-background_callback_manager = DiskcacheManager(cache)
+# Essential for initializing callbacks. Do not remove.
+import demo_callbacks
 
 # Fix Dash long callbacks crashing on macOS 10.13+ (also potentially not working
 # on other POSIX systems), caused by https://bugs.python.org/issue33725
@@ -42,6 +38,9 @@ import multiprocess
 if multiprocess.get_start_method(allow_none=True) is None:
     multiprocess.set_start_method("spawn")
 
+cache = diskcache.Cache("./cache")
+background_callback_manager = DiskcacheManager(cache)
+
 app = dash.Dash(
     __name__,
     meta_tags=[{"name": "viewport", "content": "width=device-width, initial-scale=1"}],
@@ -50,7 +49,6 @@ app = dash.Dash(
 )
 app.title = APP_TITLE
 
-server = app.server
 app.config.suppress_callback_exceptions = True
 
 # Parse debug argument
@@ -79,170 +77,14 @@ css = f"""/* Automatically generated theme settings css file, see app.py */
     --theme-secondary: {THEME_COLOR_SECONDARY};
 }}
 """
-with open("assets/custom_00_theme.css", "w") as f:
+with open("assets/__generated_theme.css", "w") as f:
     f.write(css)
 
 
-@app.callback(
-    Output({"type": "to-collapse-class", "index": MATCH}, "className"),
-    inputs=[
-        Input({"type": "collapse-trigger", "index": MATCH}, "n_clicks"),
-        State({"type": "to-collapse-class", "index": MATCH}, "className"),
-    ],
-    prevent_initial_call=True,
-)
-def toggle_left_column(collapse_trigger: int, to_collapse_class: str) -> str:
-    """Toggles a 'collapsed' class that hides and shows some aspect of the UI.
-
-    Args:
-        collapse_trigger (int): The (total) number of times a collapse button has been clicked.
-        to_collapse_class (str): Current class name of the thing to collapse, 'collapsed' if not
-            visible, empty string if visible.
-
-    Returns:
-        str: The new class name of the thing to collapse.
-    """
-
-    classes = to_collapse_class.split(" ") if to_collapse_class else []
-    if "collapsed" in classes:
-        classes.remove("collapsed")
-        return " ".join(classes)
-    return to_collapse_class + " collapsed" if to_collapse_class else "collapsed"
-
-
-@app.callback(
-    Output("input", "children"),
-    inputs=[
-        Input("slider", "value"),
-    ],
-)
-def render_initial_state(slider_value: int) -> str:
-    """Runs on load and any time the value of the slider is updated.
-        Add `prevent_initial_call=True` to skip on load runs.
-
-    Args:
-        slider_value: The value of the slider.
-
-    Returns:
-        str: The content of the input tab.
-    """
-    return f"Put demo input here. The current slider value is {slider_value}."
-
-
-class RunOptimizationReturn(NamedTuple):
-    """Return type for the ``run_optimization`` callback function."""
-
-    results: str = dash.no_update
-    problem_details_table: list = dash.no_update
-    # Add more return variables here. Return values for callback functions
-    # with many variables should be returned as a NamedTuple for clarity.
-
-
-@app.callback(
-    # The Outputs below must align with `RunOptimizationReturn`.
-    Output("results", "children"),
-    Output("problem-details", "children"),
-    background=True,
-    inputs=[
-        # The first string in the Input/State elements below must match an id in dash_html.py
-        # Remove or alter the following id's to match any changes made to dash_html.py
-        Input("run-button", "n_clicks"),
-        State("solver-type-select", "value"),
-        State("solver-time-limit", "value"),
-        State("slider", "value"),
-        State("dropdown", "value"),
-        State("checklist", "value"),
-        State("radio", "value"),
-    ],
-    running=[
-        (Output("cancel-button", "className"), "", "display-none"),  # Show/hide cancel button.
-        (Output("run-button", "className"), "display-none", ""),  # Hides run button while running.
-        (Output("results-tab", "disabled"), True, False),  # Disables results tab while running.
-        (Output("results-tab", "label"), "Loading...", "Results"),
-        (Output("tabs", "value"), "input-tab", "input-tab"),  # Switch to input tab while running.
-        (Output("run-in-progress", "data"), True, False),  # Can block certain callbacks.
-    ],
-    cancel=[Input("cancel-button", "n_clicks")],
-    prevent_initial_call=True,
-)
-def run_optimization(
-    # The parameters below must match the `Input` and `State` variables found
-    # in the `inputs` list above.
-    run_click: int,
-    solver_type: Union[SolverType, int],
-    time_limit: float,
-    slider_value: int,
-    dropdown_value: int,
-    checklist_value: list,
-    radio_value: int,
-) -> RunOptimizationReturn:
-    """Runs the optimization and updates UI accordingly.
-
-    This is the main function which is called when the ``Run Optimization`` button is clicked.
-    This function takes in all form values and runs the optimization, updates the run/cancel
-    buttons, deactivates (and reactivates) the results tab, and updates all relevant HTML
-    components.
-
-    Args:
-        run_click: The (total) number of times the run button has been clicked.
-        solver_type: Either Solver 1 (``0`` or ``SolverType.SOLVER_1``),
-            or Solver 2 (``1`` or ``SolverType.SOLVER_2``).
-        time_limit: The solver time limit.
-        slider_value: The value of the slider.
-        dropdown_value: The value of the dropdown.
-        checklist_value: A list of the values of the checklist.
-        radio_value: The value of the radio.
-
-    Returns:
-        A NamedTuple (RunOptimizationReturn) containing all outputs to be used when updating the HTML
-        template (in ``dash_html.py``). These are:
-
-            results: The results to display in the results tab.
-            problem-details: List of the table rows for the problem details table.
-    """
-
-    # Only run optimization code if this function was triggered by a click on `run-button`.
-    # Setting `Input` as exclusively `run-button` and setting `prevent_initial_call=True`
-    # also accomplishes this.
-    if run_click == 0 or ctx.triggered_id != "run-button":
-        raise PreventUpdate
-
-    if isinstance(solver_type, int):
-        solver_type = SolverType(solver_type)
-
-
-    # The following are example Input and State callback variables.
-    # This print statement is for instructional purposes and can be removed.
-    print(
-        f"The form has the following values:\n\
-        Example Slider: {slider_value}\n\
-        Example Dropdown: {dropdown_value}\n\
-        Example Checklist: {checklist_value}\n\
-        Example Radio: {radio_value}\n\
-        Solver: {solver_type}\n\
-        Time Limit: {time_limit}"
-    )
-
-    ###########################
-    ### YOUR CODE GOES HERE ###
-    ###########################
-
-    # Generates a list of table rows for the problem details table.
-    problem_details_table = generate_problem_details_table_rows(
-        solver="Solver 2" if solver_type is SolverType.SOLVER_2 else "Solver 1",
-        time_limit=time_limit,
-    )
-
-    return RunOptimizationReturn(
-        results="Put demo results here.",
-        problem_details_table=problem_details_table,
-    )
-
-
-# Imports the Dash HTML code and sets it in the app.
-# Creates the visual layout and app (see `dash_html.py`)
-set_html(app)
-
-# Run the server
 if __name__ == "__main__":
+    # Imports the Dash HTML code and sets it in the app.
+    # Creates the visual layout and app (see `demo_interface.py`)
+    app.layout = create_interface()
+
+    # Run the server
     app.run_server(debug=DEBUG)
